requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==4.2.17
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
